# openspec/changes/sdlc-visualizer/interface-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Arsitect SDLC Visualizer API
  version: 1.0.0
  description: Arsitect 可视化驾驶舱 REST API 契约（基于详细设计 20/21 模块导出）
servers:
- url: /api/v1
  description: API v1 基线
paths:
...
  /projects/{project_id}/wireframes:
    post:
      tags:
      - wireframes
      summary: Create Wireframe
      description: Create a new wireframe for a project.
      operationId: create_wireframe_api_v1_projects__project_id__wireframes_post
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WireframeCreateDTO'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WireframeResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - wireframes
      summary: List Wireframes
      description: List wireframes for a project.
      operationId: list_wireframes_api_v1_projects__project_id__wireframes_get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WireframeResponseDTO'
                title: Response List Wireframes Api V1 Projects  Project Id  Wireframes Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /wireframes/{wireframe_id}:
    get:
      tags:
      - wireframes
      summary: Get Wireframe
      description: Get a single wireframe by ID.
      operationId: get_wireframe_api_v1_wireframes__wireframe_id__get
      parameters:
      - name: wireframe_id
        in: path
        required: true
        schema:
          type: string
          title: Wireframe Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WireframeResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - wireframes
      summary: Update Wireframe
      description: Update a wireframe.
      operationId: update_wireframe_api_v1_wireframes__wireframe_id__patch
      parameters:
      - name: wireframe_id
        in: path
        required: true
        schema:
          type: string
          title: Wireframe Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WireframeUpdateDTO'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WireframeResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - wireframes
      summary: Delete Wireframe
      description: Delete a wireframe.
      operationId: delete_wireframe_api_v1_wireframes__wireframe_id__delete
      parameters:
      - name: wireframe_id
        in: path
        required: true
        schema:
          type: string
          title: Wireframe Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    WireframeCreateDTO:
      properties:
        name:
          type: string
          maxLength: 128
          minLength: 1
          title: Name
        nodes:
          type: string
          title: Nodes
          default: '[]'
        edges:
          type: string
          title: Edges
          default: '[]'
        status:
          type: string
          title: Status
          default: DRAFT
      type: object
      required:
      - name
      title: WireframeCreateDTO
      description: Create wireframe request.
    WireframeResponseDTO:
      properties:
        wireframe_id:
          type: string
          title: Wireframe Id
        project_id:
          type: string
          title: Project Id
        name:
          type: string
          title: Name
        nodes:
          type: string
          title: Nodes
        edges:
          type: string
          title: Edges
        status:
          type: string
          title: Status
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      type: object
      required:
      - wireframe_id
      - project_id
      - name
      - nodes
      - edges
      - status
      - created_at
      - updated_at
      title: WireframeResponseDTO
      description: Wireframe response.
    WireframeUpdateDTO:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 1
          - type: 'null'
          title: Name
        nodes:
          anyOf:
          - type: string
          - type: 'null'
          title: Nodes
        edges:
          anyOf:
          - type: string
          - type: 'null'
          title: Edges
        status:
          anyOf:
          - type: string
            pattern: ^(DRAFT|ACTIVE|ARCHIVED)$
          - type: 'null'
          title: Status
      type: object
      title: WireframeUpdateDTO
      description: Update wireframe request.
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: MVP 阶段为本地会话占位，P1 升级为 JWT
...